Example Peer-to-Peer Chat Program
A quick-and-easy little peer-to-peer chatting program that uses my P2P C++ library. Pretty simple... Enter in another computer's IP address, enter in a username for you to use and type a message to send. When you receive a message, the message window show's the sender's IP and their username, along with their message, of course. When you receive at least one message, you will not need to specify an IP address to send messages to. Your message will go to the last IP you received a message from.