Minor bug fixes
-Now featuring a Command Prompt-friendly version of this library, PeerToPeer_Win32.h, that uses 
CStdString as a substitute for CString, which is used in MFC Applications. I did that because
I have had trouble trying to load CStrings into console applications. This version of the library
will contain the exact updates as the regular version.

-Added EpochSeconds(), returns the current time in Epoch format, as a CString.

-Added Basic file-handling functions that return true or false if they successfully processed.

-File-handling functions are loadfile(), writefile(), appendfile() and foundinfile().

-Added an example MFC program that does Peer-to-Peer chat, used port 6969 as an example.
Type in a username, and IP to send a message to and the message itself.
When you receive a message, you don't need to type in an IP, the program will send your message
to the last IP you received a message to.